requirements.txt
celery
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.11.2
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
psycopg2==2.9.6
pyotp==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
